Move graphql to separate folder and map data from new api in resolvers
