+++ libs/api/domains/national-registry/src/index.ts
@@ -1,3 +1,1 @@
-export { FamilyMemberResolver as NationalRegistryFamilyMemberResolver } from './lib/familyMember.resolver'
 export { NationalRegistryModule } from './lib/nationalRegistry.module'
-export { UserResolver as NationalRegistryUserResolver } from './lib/user.resolver'
