+++ libs/api/domains/national-registry/src/lib/graphql/familyMember.resolver.ts
@@ -1,15 +1,16 @@
 import { UseGuards } from '@nestjs/common'
 import { Resolver, Query } from '@nestjs/graphql'
 
-import { NationalRegistryFamilyMember } from './models'
-import { NationalRegistryService } from './nationalRegistry.service'
-import { FamilyMember } from './types'
 import {
   IdsAuthGuard,
   ScopesGuard,
   CurrentUser,
   User as AuthUser,
 } from '@island.is/auth-nest-tools'
+
+import { NationalRegistryFamilyMember } from './models'
+import { NationalRegistryService } from '../nationalRegistry.service'
+import { FamilyMember } from '../types'
 
 @UseGuards(IdsAuthGuard, ScopesGuard)
 @Resolver(() => NationalRegistryFamilyMember)
