+++ libs/api/domains/national-registry/src/lib/graphql/index.ts
@@ -1,0 +1,3 @@
+export { FamilyMemberResolver } from './familyMember.resolver'
+export { UserResolver } from './user.resolver'
+export * from './scalars'
