+++ libs/api/domains/national-registry/src/lib/graphql/models/banMarking.model.ts
