+++ libs/api/domains/national-registry/src/lib/graphql/models/familyMember.model.ts
@@ -1,5 +1,6 @@
 import { Field, ObjectType, ID } from '@nestjs/graphql'
-import { Gender, MaritalStatus, FamilyRelation } from '../types'
+
+import { Gender, MaritalStatus, FamilyRelation } from '../../types'
 
 @ObjectType()
 export class NationalRegistryFamilyMember {
@@ -11,12 +12,6 @@ export class NationalRegistryFamilyMember {
 
   @Field(() => Gender, { nullable: true })
   gender?: Gender
-
-  @Field(() => MaritalStatus, { nullable: true })
-  maritalStatus?: MaritalStatus
-
-  @Field(() => String)
-  address!: string
 
   @Field(() => String)
   familyRelation!: FamilyRelation
