+++ libs/api/domains/national-registry/src/lib/graphql/models/index.ts
