+++ libs/api/domains/national-registry/src/lib/graphql/models/user.model.ts
@@ -1,5 +1,6 @@
 import { Field, ObjectType, ID } from '@nestjs/graphql'
-import { Gender, MaritalStatus } from '../types'
+
+import { Gender, MaritalStatus } from '../../types'
 import { BanMarking } from './banMarking.model'
 
 @ObjectType()
